requirements.txt
# Web framework & server
Flask==2.3.3
gunicorn==21.2.0

# Auth & DB
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9

# Payments
stripe==5.8.0

# Env loader
python-dotenv==1.0.1

# Flask deps pinned for stability
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
